issue483: move the logic from subgraph_slot.py
