+++ mypy.ini
@@ -44,6 +44,9 @@ ignore_missing_imports = True
 [mypy-pdr_backend.predictoor.examples.*]
 ignore_missing_imports = True
 
+[mypy-polars.*]
+ignore_missing_imports = True
+
 [mypy-pylab.*]
 ignore_missing_imports = True
